scripts/pyproject.toml
[tool.poetry]
name = "scripts"
version = "0.1.0"
description = ""
authors = ["arfrie22 <43021241+arfrie22@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
